requirements.txt
numpy
opencv_python>=4.1.1.26
h5py>=2.10.0
ConfigArgParse>=0.15.1
imgaug>=0.3.0
matplotlib>=2.2.4
keras==2.3.0
scipy>=1.3.2
pip>=19.2.2
tensorflow==1.15.2
pandas>=0.25.1
scikit-image
ipython>=7.13.0
packaging>=20.3
Pillow>=7.1.1
jupyter
jupyterlab
segmentation-models==0.2.1
libtiff
